+++ packages/@ember/object/package.json
@@ -9,7 +9,6 @@
     "./internals": "./internals.ts",
     "./evented": "./evented.ts",
     "./core": "./core.ts",
-    "./observable": "./observable.ts",
     "./events": "./events.ts",
     "./computed": "./computed.ts",
     "./compat": "./compat.ts",
