+++ packages/ember/tests/reexports_test.js
@@ -80,7 +80,6 @@ import * as test24 from '@ember/object/core';
 import * as test26 from '@ember/object/events';
 import * as test27 from '@ember/object/internals';
 import * as test28 from '@ember/object/mixin';
-import * as test29 from '@ember/object/observable';
 import * as test30 from '@ember/object/observers';
 import * as test33 from '@ember/routing/hash-location';
 import * as test34 from '@ember/routing/history-location';
@@ -206,7 +205,6 @@ let allExports = [
   ['cacheFor', '@ember/object/internals', 'cacheFor', test27],
   ['guidFor', '@ember/object/internals', 'guidFor', test27],
   ['Mixin', '@ember/object/mixin', 'default', test28],
-  ['Observable', '@ember/object/observable', 'default', test29],
   ['addObserver', '@ember/object/observers', 'addObserver', test30],
   ['removeObserver', '@ember/object/observers', 'removeObserver', test30],
   ['HashLocation', '@ember/routing/hash-location', 'default', test33],
