+++ tests/docs/expected.js
@@ -522,7 +522,6 @@ module.exports = {
     'Mixin',
     'Namespace',
     'NoneLocation',
-    'Observable',
     'Owner',
     'Promise',
     'RegisterOptions',
@@ -555,7 +554,6 @@ module.exports = {
     '@ember/object',
     '@ember/object/core',
     '@ember/object/mixin',
-    '@ember/object/observable',
     '@ember/owner',
     '@ember/renderer',
     '@ember/routing',
